requirements.txt
Flask==3.0.0
requests==2.31.0
python-dotenv==1.0.0
PyGithub==2.1.1
openai==1.3.0
pylint==3.0.0
radon==6.0.1
coverage==7.3.0
beautifulsoup4==4.12.2
lxml==4.9.3
pytest==7.4.3